Break study pane into smaller components (#415)
