+++ src/components/app.js
@@ -17,15 +17,17 @@ import {
     useLocation,
 } from 'react-router-dom';
 
-import StudyPane, { StudyView } from './study-pane';
+import { StudyView } from './study-pane';
 import {
     changeDisplayedColumns,
     resetLoadflowNotif,
     resetSANotif,
     selectCenterLabelState,
+    selectComponentLibrary,
     selectComputedLanguage,
     selectDiagonalLabelState,
     selectDisplayOverloadTableState,
+    selectFavoriteContingencyLists,
     selectLanguage,
     selectLineFlowAlertThreshold,
     selectLineFlowColorMode,
@@ -33,10 +35,8 @@ import {
     selectLineFullPathState,
     selectLineParallelPathState,
     selectSubstationLayout,
-    selectComponentLibrary,
     selectTheme,
     selectUseName,
-    selectFavoriteContingencyLists,
 } from '../redux/actions';
 
 import {
@@ -80,6 +80,7 @@ import { getComputedLanguage } from '../utils/language';
 import AppTopBar from './app-top-bar';
 import { useSnackbar } from 'notistack';
 import { displayErrorMessageWithSnackbar, useIntlRef } from '../utils/messages';
+import { StudyContainer } from './study-container';
 
 const noUserManager = { instance: null, error: null };
 
@@ -334,10 +335,12 @@ const App = () => {
     }, []);
 
     // if result tab is displayed, clean badge
-    if (STUDY_VIEWS[tabIndex] === StudyView.RESULTS) {
-        dispatch(resetSANotif());
-        dispatch(resetLoadflowNotif());
-    }
+    useEffect(() => {
+        if (STUDY_VIEWS[tabIndex] === StudyView.RESULTS) {
+            dispatch(resetSANotif());
+            dispatch(resetLoadflowNotif());
+        }
+    }, [tabIndex, dispatch]);
 
     return (
         <div
@@ -372,7 +375,7 @@ const App = () => {
                 {user !== null ? (
                     <Switch>
                         <Route exact path="/studies/:studyUuid">
-                            <StudyPane
+                            <StudyContainer
                                 view={STUDY_VIEWS[tabIndex]}
                                 onChangeTab={onChangeTab}
                             />
