+++ src/components/menus/equipment-menu.js
